fix: pre-release bug sweep — 10 confirmed bugs fixed
Critical:
- Stacking disk-path used wrong callback (_on_alignment_done instead of
  _on_path_alignment_done), causing AttributeError on Path objects
- Continuum subtraction returned 2D mono array for color images; now
  broadcasts result back to 3D to preserve channel count
- FITS normalization checked dtype after astype(float32), making integer
  branches dead code; check dtype before converting
- GPU warp was only scaling translation components to normalized coords,
  causing ~10px shear on non-square images; full matrix now converted

Serious:
- Unsharp mask, median filter, vignette correction, CA correction, and
  stretch all ran synchronously on the main thread; moved to _start_worker

Minor:
- AI denoise/sharpen had messy dead code inside the None-check guard;
  replaced with proper log call
- Help menu linked to wrong GitHub repo (cosmica-app vs majmichu1)
- Calibration import hardcoded FrameType.FLAT for all frames; now uses
  _guess_frame_type to auto-detect bias/dark/flat from filename
- Phantom self._aligned_lights = [] attribute removed (del did the actual
  freeing; the instance attr was never used)
- Duplicate @pyqtSlot() decorator on _on_start_crop_draw removed

+++ cosmica/core/image_io.py
@@ -187,17 +187,14 @@ def auto_stretch_for_display_ref(img: np.ndarray, ref: np.ndarray) -> np.ndarray
 
 def _normalize_fits_data(data: np.ndarray) -> np.ndarray:
     """Normalize FITS data to float32 in [0, 1]."""
-    data = data.astype(np.float32)
     if data.dtype.kind == "u":
-        # unsigned integer
         max_val = np.iinfo(data.dtype).max
-        return data / max_val
+        return (data / max_val).astype(np.float32)
     elif data.dtype.kind == "i":
-        # signed integer — offset
         info = np.iinfo(data.dtype)
-        return (data.astype(np.float32) - info.min) / (info.max - info.min)
+        return ((data.astype(np.float32) - info.min) / (info.max - info.min))
     else:
-        # float — check range
+        data = data.astype(np.float32)
         dmin, dmax = float(np.min(data)), float(np.max(data))
         if dmax <= 1.0 and dmin >= 0.0:
             return data
